tdf#56258 tdf#106988 sw numbering ivx: make style useable
In the current state, the style was basically unusable.
The first two levels had numbering at the same indent,
the third level started with a c),
and the fourth level started earlier than the 3rd level.

I took my formatting queues from Numbering IVX.
These right-aligned roman-numerals can grow very wide
when getting to number 7 and 8, so I designed the
style with a wide indent theme.

Now the first two levels have the same width as
ROMAN_UPPER. Since this style is a mixed bag,
I changed the second level to use ROMAN_UPPER,
which keeps the roman theme, but fits better with
the following character and bullet sublevels.
The third level switches back to left-aligned
with similar indents to CHAR_LOWER_LETTER,
(and starts with letter "a" - good grief)
while the bullet levels are a nice, breezy half-indent width
(which is 50%-100% wider than the bullet styles).

Change-Id: Ie3ac386c360e804dde6c295dd39a3c02d3ba9b79
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144841
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Vernon, Stuart Foote <vsfoote@libreoffice.org>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150649
Tested-by: Justin Luth <jluth@mail.com>