sw, lok: implement a getCommandValues(Fields)
There was no LOK API to get a list of all fields of a given type where
the name matches a certain prefix.

This is useful in case the API cilent wants to know what previously
inserted refmarks were deleted by the user as part of deleting text
content.

Add a new getCommandValues(".uno:Fields") that returns the names of
matching refmarks. Do not return the refmark text, assuming that would
be updated by the API client anyway.

In practice this is needed by Zotero in case it wants to model its
citations with refmarks.

Change-Id: Ie469253891896aa8ab00d434c9ab116adbe3864b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/144985
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins