tdf#56258 svx: use last defined locale-outline for remaining levels
When applying an outline numbering choice
(one of the 8 predefined per-locale definitions)
the toolbar/sidebar would only affect the spacing
of 0 - 5 of the defined locale-levels,
and only changed the num type of the remainder.

Well, that can easily look bad if the indenting and spacing
is only partially modified.
[The Bullets and Numbering dialog outline grid
 does not affect spacing at all.
 That can be better or much worse, depending on the situation.]

This "Outline Format" button is normally turned off
on Writer's Formatting toolbar. You need to go to
Tools - Customize - Toolbars to turn it on.
It only seems to work for Writer - it was grayed out/disabled
in Draw even when I had it show up in the toolbar.

So almost nobody have ever seen or used this I expect.
However, I think it is important to get this working,
because I want to use this grid-choice code path
to be usable for the B&N dialog as well,
although that could upset people who already have
their spacing right and just want to change the digits.

In any case, this nbdtimg DOES already change some
spacing, so it should be either all or none.

Change-Id: Ib54b6ffe3497f09f0c57ff2db1ec2a946f692fc5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150660
Tested-by: Justin Luth <jluth@mail.com>
Reviewed-by: Justin Luth <jluth@mail.com>