sw lok: expose field type & command of fieldmark under cursor
It was possible to get the properties of all fieldmarks, but you could
not get the properties of the fieldmark under the current cursor.

Getting the properties of the current fieldmark is useful for Zotero: if
we already have a citation and want to insert one more, then we should
turn the current citation into a citation cluster.

Fix the problem by adding API similar to what commit
24219cc1e9829f82a533667aef0f51b6a7df6fc2 (sw lok, .uno:TextFormFields:
expose field code of fieldmarks, 2022-11-25), did: but here we return
properties of the innermost fieldmark, not all fieldmarks.

Also introduce a ITiledRenderable::supportsCommandValues(), this way
next time a command value getter is added in sw/sc/sd, no need to touch
desktop/ anymore.

(cherry picked from commit bb20dee2ef1b0804065e1cda2c834d257fdd90ed)

Conflicts:
	sw/source/uibase/uno/loktxdoc.cxx

Change-Id: I7f1a7064307034a18527ab5e985d2eac56807cb7