sw update of refmarks: fix handling of ignored refmarks
As pointed out at
<https://gerrit.libreoffice.org/c/core/+/145036/1#message-165c0282ace92160592c896c6867095d3558ee96>,
it's not correct that UpdateFieldConents() uses a single index into both
the provided fields/refmarks array and into the document's all refmarks
array.

We need a new index that counts the input fields/refmarks we got and
which is not incremented for ignored refmarks.

Extend the testcase to have 2 paragraphs in the document: the first
paragraph has a refmark that is to be ignored and the second para has a
refmark that is interesting to us. This makes the test fail.

Then fix up UpdateFieldConents() to properly count how we iterate
through all refmarks and the provided refmarks in parallel.

(cherry picked from commit 471804e251b4e15b37a10920bd4b88b40f97b227)

Change-Id: I1fabf5b29beeabe9e8a126f1662ca8830e5c4e20