sw: rework ThemePanel, theme color ValueSet drawing as custom draw
Rework the ThemePanel to only work with theme colors, as fonts
are uing an old mock implemntation of themes. To do this properly
it is needed to remove the implementation for now.

The theme colors entries were rendered as a bitmap and then shown
in the ValueSet. This has the problem that it doesn't look sharp
on a HiDPI screen, so replace that with a ThemeColorValueSet, which
uses custom draw to render the theme color entries directly.

The ThemeColorValueSet was added to th svx component as it will be
used in other components as well.

Change-Id: I1a727ef4cf4be4e215db57ac33f571f43aaddc15
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145087
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit 8bc22ed6899bcbafc3020f0af6c939019506a5fd)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145853
Tested-by: Tomaž Vajngerl <quikee@gmail.com>