tdf#56258 i18npool en-US: change outline order to match MLA style
Changed I.A.i.a) -> I.A.1.a.i.

The overwhelming result of a google search for
"official/correct outline list format"
indicates that the third level should be a 1, not an i.

MLA Handbook (according to wikipedia) uses the order
I.A.1.a.i. (a) (1) i)

Chicago Mode of Style is usually interpreted as
I.A.1.a) or I.A.1.a. with varying lower sublevels/suffixes.

Wikipedia also suggest the logical order of
I.A.1.a.i.I) A) 1) a) i) (I) (A) (1) (a) (i)

Change-Id: I7fcc8a1ee727d6cddca4bf0c6a8461538769cf5d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145147
Tested-by: Justin Luth <jluth@mail.com>
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/150664