sw: .uno:TextFormField: add new Wrapper parameter
Currently all fieldmarks are inserted into the document body
unconditionally when this UNO command is dispatched.

Inserting at the current cursor position makes sense, but some citation
styles want to insert the actual citation as footnotes, and only have
the footnote anchor at the cursor position.

Fix the problem by adding a new Wrapper parameter to this UNO command:
currently the only interesting value it may have is Footnote, if this is
specified then first we insert a footnote and the footnote content will
host the fieldmark, not the original body text.

The same will be wanted for endnotes as well, but that's not yet done in
this commit.

(cherry picked from commit ceea8f3924f26d5f10adc41b9ea587c77c2fda74)

Change-Id: I5c96c7dc9ddaace09b1dbc21b8f12005a2934d04