sw: lok: use redline author for saved author fields(2)
Without the fix author fields will expand to "Unknown author" when
loading files which have author fields in them. But only update the
fields when the first view joins and not for later view joins.

online.git does not expect document to be modified just after load, so
to avoid this situation clear the document modified flag just after
field update.

Change-Id: I19253f4c871e2c023d206d2d83b6d694ed5933d1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145654
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>