tdf#153083 writerfilter: import locale-dependent TOC \t style names
The bugdoc contains this style:

  <w:style w:type="paragraph" w:styleId="IntensivesZitat">
    <w:name w:val="Intense Quote"/>
    <w:basedOn w:val="Standard"/>
    <w:next w:val="Standard"/>
    <w:link w:val="IntensivesZitatZchn"/>
    ...

which is referred to by:

  TOC \o "1-3" \h \z \t "Intensives Zitat;3;Custom1;3;_MyStyle0;3"

Word in an "en" locale is unable to match the "Intensives Zitat" in the
TOC field with the style "Intense Quote", which is a built-in style
in Word (no equivalent in Writer).

At first glance nothing in styles.xml matches the localised built-in
style name in the TOC field.

But it looks like the w:styleId value is somehow generated from the
localised style name by omitting certain characters like SPACE and
non-ASCII letters.

Change-Id: I2050f7cf7f8d80bee1f667ee53b7f9981bbf7b49
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145745
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
(cherry picked from commit ecbad22fdf81c6f072b6c9f9c16dbba47fe4748c)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/145721
Tested-by: Michael Stahl <michael.stahl@allotropia.de>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>