xmloff: export Theme for ODT (Writer) documents
Exports the Theme set on the one and only SdrPage for ODT (Writer)
documents. The theme is exported as a sub-element of "style".
This differs to ODP (Impress) documents, which export the Theme
as part of the master-page element.

Currently the code is duplicated, because the theme can't yet be
accessed inside xmloff in a non-conflicting way (creating conflicts
between modules), which will be fixed in the future.

Change-Id: I0e440d14724b49e7d86e9deabc6615a91e8cd31a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146172
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit 6d1413da02602992b42b62bda9a26ffeb774b603)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146227
Tested-by: Tomaž Vajngerl <quikee@gmail.com>