sw: add an SwFlyAtContentFrame ctor that creates a follow fly frame
This is meant to be a called by a future SwFrame::GetNextFlyLeaf()
(similar to the existing SwFrame::GetNextSctLeaf()), which still has to
be added.

What an anchor should be for a follow fly frame is a bit unclear to me,
let's go with an explicit nullptr for now and we can clarify that later.

Towards an initial layout for multi-page fly frames.

(cherry picked from commit a6b64d02843d186485ff5c82106e6f9268a539c7)

Change-Id: I2429f47aef90dfe61340bfe244353dcefeb4ef1f