tdf#151548 sw content controls: actually gotoFormatContentControl
This is a squashed commit, including Caolan's
91fa7cc3dc124877f810599fe3c315cda6d03df6
    cid#1520798 Use after free

The cursor kindof needs to change when going to a content control,
so don't return early just because there will be no selection.
Instead, set the cursor at the beginning of the control,
and simply skip the selection part.

Without this, it was not possible for the keyboard navigation
to jump to rich/plain text or comboboxes that were no longer
showing placeholder content.

Change-Id: I055ece1d296c6cdf70d6bc7c1df7797bf09532c9
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146266
Tested-by: Jenkins
Reviewed-by: Justin Luth <jluth@mail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146663
Tested-by: Justin Luth <jluth@mail.com>