sw: implement SwFrame::GetPrevFlyLeaf()
This is much easier than the "next" case, since here we can nullptr when
there is no previous leaf instead of creating one.

With this, if the MoveBwd() call in SwContentFrame::MakeAll() is
disabled, some simple fly frame can be split across 2 pages if it's
positioned in a way that the rectangle of the fly would not fit into the
body frame without changing the position, in SW_FORCE_FLY_SPLIT=1 mode.

Towards an initial layout for multi-page fly frames.

(cherry picked from commit 27fbab13557a75b5402c11a1697541edc124116a)

Change-Id: I3b6ba85ad66117b5595a113d688e2fcb97bcf745