tdf#153083 writerfilter,sw: DOCX locale-dependent TOC \t style names, 3
The import crashed on fdo85740-1.docx with an unhandled exception.

This is because it mapped the TOC style "Table of Figures" from "table
of figures", but the ApplyStyleSheetsImpl() actually uses the converted
name "Drawing" instead, so "Table of Figures" didn't exist.

(regression from commit ca71482237d31703454062b8b2f544a8bacd2831)

It turns out that once that is fixed, the DOCX export needs to convert
"Drawing" back to "Table of Figures" in the TOC field.

-- but apparently not on the libreoffice-7-5 branch, as here the style
names are written unconverted; that is only needed since commit
319e2e0cbc6a1d25abd01fb27d6250cee825c072 so omit export changes here.

Change-Id: I4858c79dd74154b229b7568610c0b8ba7b3e2b6a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146610
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>
(cherry picked from commit 345f23474845da0defa4a9276228f54b5ded5736)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/146639
Reviewed-by: Caolán McNamara <caolanm@redhat.com>