sw floattable: handle table-in-fly in SwFrame::GetNextFlyLeaf()
Trying to lay out a split fly that contains a table resulted in a stack
overflow.

The reason for this was that SwObjectFormatter::FormatObjsAtFrame_() has
a loop that will format all objects of the current fly frame, but we
managed to anchor the follow fly into itself.

Fix the problem by improving SwFrame::GetNextFlyLeaf(), somewhat based
on how SwFrame::GetNextSctLeaf() has special cases for tables. This way
once we split the fly frame, we'll try to move the follow anchor frame
to the next page's body frame, and not to a child of the follow fly
itself.

Also add a first floattable testcase, we can already assert that the
table is split correctly. Do this in a separate suite for now, since the
pool's default SwFormatFlySplit is only created once, so
SwFormatFlySplit::SetForce(false) doesn't have the wanted effect. The
anchor's text is still on both pages, should be on page 2 only.

(cherry picked from commit 995198bfff4ae8abaf2129fe99d9f8ef899a4f25)

Change-Id: Ie2ce75dbace5d9716008351aedb6a8989036badb