sw floattable: improve handling of split flys in SwTextFrame::FormatEmpty()
The bugdoc had a non-empty paragraph after the floating table, we
painted that text on both page 1 and page 2, while only page 2 is
expected.

We want that paragraph to be on both pages, but we want to split it
before the first char, so apart from providing a place to serve as an
anchor, it should be invisible on page 1.

Fix the problem by extending SwTextFrame::FormatEmpty() to consider the
master anchor text frame as empty in case it has no text, just
intersects with a fly frame.

Also improve SwTextFrame::HasNonLastSplitFlyDrawObj() a bit, so it
returns early when the own and the follow offset is not the same. These
are cheap checks and some callers already did it before iterating over
all the draw objects of the text frame, but other callers didn't do
this. Now we check for this consistently, everywhere.

(cherry picked from commit 1cf13b8b11e7c122631394e308b66b46ffaae718)

Change-Id: I4b399ae3f9e9f364c61d977359f6c9b16c648629