tdf#151968: Fix vertical position of RTL spelling wavy line
The code was guessing orientation based on the start and end points and
mistakenly considered Arabic text to be 180° rotated which ended up
raising the wavy line above baseline and covering the RTL text.

Use font orientation instead of guessing it. Caching wavy line seems
broken with RTL text as well (it was skipped for RTL before because of
the wrong guessed orientation, so probably never worked), so we skip it
explicitly as well for now.

Change-Id: I4b0f5c4d6be2c6e3d33ea79b917b14927374acfa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147488
Tested-by: Jenkins
Reviewed-by: خالد حسني <khaled@aliftype.com>
(cherry picked from commit 5899b27e71430e490c2d3a6b87ae52c10f383ba7)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/147441
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>